+++ lib/common_generators.dart
@@ -3,4 +3,3 @@ import 'operations.dart';
 
 part 'src/common_generators/operations_to_code_converter.dart';
 part 'src/common_generators/utils.dart';
-