+++ lib/operation_transformers.dart
@@ -1,3 +1,4 @@
+import 'operation_optimizers.dart';
 import 'operations.dart';
 
 part 'src/operation_transformers/operation_initializer.dart';